move Postgresql tests from test_api.py to PG provider tests
